requirements.txt
streamlit
torch
torchvision
segmentation-models-pytorch
opencv-python-headless
albumentations
Pillow
numpy
